V0.10.3 - Bug fix Reverted to old method of preventing locking
This may or may not work, requires further Alpha testing at work and home
to determin if it is a code problem or PC setting problem. Could not
reproduce the bug where the time until home counter would be wrong on
Friday's. Should fix loading mouse over bug (not documented).