Version 0.10.5 Alpha Added working keep unlocked method
Using code from mouse jiggler and Win32 user interaction timer, the
program now checks every 55s to see if the computer has been interacted
with in the last 50s. If it hasn't it will move the mouse +1 pixel in x
and y direction, which resets the interaction timer.

This can be disabled in the settings menu if needed.